LCD driver issues: add MIOS pin for ST77789 input/output - useless; display pages bugs: changes pages order; display optimization: only changes for channel, btp data. maintain shadow data (text, number, buttons) for channel, bpt and screen; pages display state machine adjustments - correction ; list test seklection button (up, down) operation

+++ bootloader/hostApplication/burnUBA6/SREC/srec.c
@@ -61,11 +61,11 @@ int extract_srec_line(char* line, int line_len, char* srec, int srec_len)
     checksum += (addr >> 8) & 0xFF;
     checksum += addr & 0xFF;
 
-    if ((addr - ETX_APP_FLASH_ADDR) > srec_len)
+    if ((int)(addr - ETX_APP_FLASH_ADDR) > srec_len)
     {
         //printf ("pad with 0xFF: adr %x, srec_len %x, pad %d\n", addr - ETX_APP_FLASH_ADDR, srec_len, ((addr - ETX_APP_FLASH_ADDR) - srec_len));
         //pad with 0xFF
-        for (int i = 0; i < ((addr - ETX_APP_FLASH_ADDR) - srec_len); i++)
+        for (int i = 0; i < (int)((addr - ETX_APP_FLASH_ADDR) - srec_len); i++)
         {
             srec[srec_len+i] = 0xFF;
         }
